+++ .gitignore
@@ -1,3 +1,4 @@
 testing
 config/db_config.php
 vendor
+
